third_party/blink/web_tests/fast/dom/shadow/multiple-host-pseudos-in-compound-expected.txt

Multiple :host(-context) pseudos in same compound.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

Should be green
